Correct repsonsible class for two methods
formatFromFileToConsole in HelperFunctions

printChirpsFromFile in UserInterface
