+++ resources/chirp_cli_db.csv
@@ -1,0 +1,10 @@
+Author,Message,Timestamp
+ropf,"Hello, BDSA students!",1690891760
+adho,"Welcome to the course!",1690978778
+adho,"I hope you had a good summer.",1690979858
+ropf,"Cheeping cheeps on Chirp :)",1690981487
+andrebirk,"helllooooo",1725104241
+andrebirk,"helllooooo",1725104248
+andrebirk,"helllooooo",1725104256
+andrebirk,"hi",1725105316
+peterolufsen,"Aksel",1725199940
